Add Swagger-Compatible Docstrings to Absences Routes
This commit introduces detailed docstrings for the absences-related routes in our Flask application. These docstrings are formatted to be compatible with Swagger, providing clear and structured documentation for each API endpoint.

Key Changes:
- Updated docstrings for GET, PUT, POST, and DELETE routes in the absences blueprint.
- Each docstring now includes tags, parameters, response descriptions, and example responses.
- Enhanced error handling documentation, specifying potential HTTP status codes like 404 (Not Found) and 500 (Internal Server Error).

These changes aim to improve the API's usability and accessibility, making it easier for developers to understand and interact with the absences management functionality.

+++ controllers/Absence_controller.py
@@ -62,22 +62,22 @@ def get_all_absences():
 @absences_bp.route('/absences/student/<int:id_student>', methods=['GET'])
 def get_student_absences(id_student):
     """
-    Récupère toutes les absences d'un étudiant via son ID.
+    Récupère toutes les absences d'un étudiant spécifié par son ID.
 
     ---
     tags:
       - Absences
     parameters:
-      - name: id_classroom
+      - name: id_student
         in: path
-        description: L'identifiant unique de l'étudiant pour qui on veux récupérer les abscenses.
-        required: true
         type: integer
+        required: true
+        description: L'identifiant unique de l'étudiant pour lequel récupérer les absences.
       - name: justified
         in: query
         type: integer
         required: false
-        description: 1 pour les absences justifiées, 0 pour les absences non justifiées.
+        description: 1 pour les absences justifiées, 0 pour les absences non justifiées. Laisser vide pour toutes les absences.
       - name: output_format
         in: query
         type: string
@@ -86,18 +86,20 @@ def get_student_absences(id_student):
         description: Format de sortie des données ('DTO' ou 'model').
     responses:
       200:
-        description: Liste des absences récupérées depuis la base de données.
+        description: Liste des absences de l'étudiant récupérées depuis la base de données.
         examples:
           application/json: [
             {
               "id": 1,
               "student_id": 123,
               "date": "2023-01-01",
-              "justified": true
+              "justified": true,
               # ... autres champs
             },
             # ... autres absences
           ]
+      404:
+        description: L'étudiant spécifié n'existe pas.
       500:
         description: Erreur serveur en cas de problème lors de la récupération des absences.
     """
@@ -120,26 +122,44 @@ def get_student_absences(id_student):
 @absences_bp.route('/absences/student/<int:id_student>/course/<int:id_course>', methods=['PUT'])
 def update_student_course_absence(id_student, id_course):
     """
-        Modifie une absence d'un étudiant pour un cours donné.
+    Modifie une absence d'un étudiant pour un cours donné.
 
-        Params:
-            id_student (int): ID de l'étudiant.
-            id_course (int): ID du cours.
-
-        JSON Request Body:
-            datas (dict): Données de l'absence à mettre à jour.
-                - reason (str): Raison de l'absence.
-                - justify (str): Justification de l'absence.
-
-        Returns:
-            JSON: Message de confirmation de la mise à jour de l'absence.
-
-        Raises:
-            404 (Not Found): L'absence ou l'étudiant spécifié n'existe pas.
-            400 (Bad Request): Données JSON manquantes ou incorrectes.
-            500 (Internal Server Error): Erreur serveur en cas de problème lors de la mise à jour de l'absence.
-        """
-
+    ---
+    tags:
+      - Absences
+    parameters:
+      - name: id_student
+        in: path
+        type: integer
+        required: true
+        description: ID de l'étudiant.
+      - name: id_course
+        in: path
+        type: integer
+        required: true
+        description: ID du cours.
+      - in: body
+        name: datas
+        required: true
+        schema:
+          type: object
+          properties:
+            reason:
+              type: string
+              description: Raison de l'absence.
+            justify:
+              type: string
+              description: Justification de l'absence.
+    responses:
+      200:
+        description: Confirmation de la mise à jour de l'absence.
+      400:
+        description: Données JSON manquantes ou incorrectes.
+      404:
+        description: L'absence ou l'étudiant spécifié n'existe pas.
+      500:
+        description: Erreur serveur lors de la mise à jour de l'absence.
+    """
     json_data = request.json
     if not json_data or 'datas' not in json_data:
         return jsonify({"message": "Données manquantes"}), 400
@@ -173,22 +193,41 @@ def add_student_course_absence(id_student, id_course):
     """
     Ajoute une absence d'un étudiant pour un cours donné.
 
-    Params:
-        id_student (int): ID de l'étudiant.
-        id_course (int): ID du cours.
-
-    JSON Request Body:
-        datas (dict): Données de l'absence à ajouter.
-            - reason (str): Raison de l'absence.
-            - justify (str): Justification de l'absence.
-
-    Returns:
-        JSON: Message de confirmation de l'ajout de l'absence.
-
-    Raises:
-        404 (Not Found): Le cours ou l'étudiant spécifié n'existe pas.
-        400 (Bad Request): Données JSON manquantes ou incorrectes.
-        500 (Internal Server Error): Erreur serveur en cas de problème lors de l'ajout de l'absence.
+    ---
+    tags:
+      - Absences
+    parameters:
+      - name: id_student
+        in: path
+        type: integer
+        required: true
+        description: ID de l'étudiant.
+      - name: id_course
+        in: path
+        type: integer
+        required: true
+        description: ID du cours.
+      - in: body
+        name: datas
+        required: true
+        schema:
+          type: object
+          properties:
+            reason:
+              type: string
+              description: Raison de l'absence.
+            justify:
+              type: string
+              description: Justification de l'absence.
+    responses:
+      201:
+        description: Confirmation de l'ajout de l'absence.
+      400:
+        description: Données JSON manquantes ou incorrectes.
+      404:
+        description: Le cours ou l'étudiant spécifié n'existe pas.
+      500:
+        description: Erreur serveur lors de l'ajout de l'absence.
     """
     try:
         json_data = request.json
@@ -229,33 +268,46 @@ def add_student_course_absence(id_student, id_course):
 @absences_bp.route('/absences/student/<int:id_student>/course/<int:id_course>/delete', methods=['DELETE'])
 def delete_student_course_absence(id_student, id_course):
     """
-    Supprime une absence d'un étudiant pour un cours donné.
+    Supprime une absence spécifique d'un étudiant pour un cours donné.
 
-    Params:
-        id_student (int): ID de l'étudiant.
-        id_course (int): ID du cours.
-
-    Returns:
-        JSON: Message de confirmation de la suppression de l'absence.
-
-    Raises:
-        404 (Not Found): L'absence ou l'étudiant spécifié n'existe pas.
-        500 (Internal Server Error): Erreur serveur en cas de problème lors de la suppression de l'absence.
+    ---
+    tags:
+      - Absences
+    parameters:
+      - name: id_student
+        in: path
+        type: integer
+        required: true
+        description: L'identifiant unique de l'étudiant concerné.
+      - name: id_course
+        in: path
+        type: integer
+        required: true
+        description: L'identifiant unique du cours concerné.
+    responses:
+      200:
+        description: Confirmation de la suppression de l'absence.
+        examples:
+          application/json: 
+            {"message": "Absence supprimée avec succès."}
+      404:
+        description: L'étudiant ou le cours spécifié n'existe pas, ou l'absence n'a pas été trouvée.
+      500:
+        description: Erreur serveur en cas de problème lors de la suppression de l'absence.
     """
     # Instanciation du service d'absences
     absences_service = AbsencesService()
 
-    # Vérification que l'étudiant existe
+    # Vérification de l'existence de l'étudiant et du cours
     if not connect_pg.does_entry_exist("Students", id_student):
         return jsonify({"message": "L'étudiant spécifié n'existe pas."}), 404
 
-    # Utilisation du service pour supprimer l'absence
+    if not connect_pg.does_entry_exist("Courses", id_course):
+        return jsonify({"message": "Le cours spécifié n'existe pas."}), 404
+
+    # Tentative de suppression de l'absence
     try:
         message = absences_service.delete_student_course_absence({"id_student": id_student, "id_course": id_course})
-        if "supprimée" in message:
-            return jsonify({"message": message}), 200
-        else:
-            return jsonify({"message": message}), 404
+        return jsonify({"message": message}), 200
     except Exception as e:
         return jsonify({"message": str(e)}), 500
-
